working towards a better Python-based profiling interaction.
The pandas library is a stable and well-used module that enables
reading and interacting with large datasets in Python easy
and quite fast. Generate a large profile (100 MB) with PaRSEC
and run the included py_dbpreader_test.py script over it if you're curious.

This will soon supersede and replace the existing Python profile
data structure. Old scripts (of which most are mine) will need to
be updated to use the new format. The effort will be well worth it.